Added bubble stabilization to Hexahedra and Quadrilateral class
Added function to compute bubble function and gradient in hexahedra.
Added function to compute bubbles in quadrilateral.
Started implementation of general assembly loop. First testing a loop
for Poromechanics. If successful, the idea is to replicate the same
concept to the mortar assembly, which is expensive and repeated several
times in the code.