Fixed bug in elementBased algorithm
A bug in the gp projection with element-based algorithm ocurred.
The cause was an error where accessing the matrix of slave basis
function with the logical array marking the gauss points to be
projected. Now this array is fixed in size for each slave element, and
an additional array flagProj is used to access the slave (and
multiplier) basis function matrix and the vector of weighed jacobian
determinant.